+++ src/core/context.ts
@@ -166,6 +166,7 @@ import { Rule, RuleSystem } from "../math/ai/rulesystem";
 import { StateMachine } from "../math/ai/statemachine";
 import { clamp } from "../math/clamp";
 import { Color, hsl2rgb, rgb } from "../math/color";
+import { buildConvexHull } from "../math/convexhull";
 import { easings } from "../math/easings";
 import { gjkShapeIntersection, gjkShapeIntersects } from "../math/gjk";
 import { lerp } from "../math/lerp";
@@ -228,7 +229,8 @@ import {
     vec2,
     wave,
 } from "../math/math";
-import { floodFill } from "../math/navigation";
+import { buildConnectivityMap, floodFill } from "../math/navigation";
+import { NavGrid } from "../math/navigationgrid";
 import { NavMesh } from "../math/navigationmesh";
 import {
     createCogPolygon,
@@ -567,6 +569,7 @@ export const createContext = (
         createStarPolygon,
         createCogPolygon,
         floodFill,
+        buildConnectivityMap,
         easingSteps,
         easingLinear,
         easingCubicBezier,
@@ -584,7 +587,9 @@ export const createContext = (
         gjkShapeIntersection,
         isConvex,
         triangulate,
+        buildConvexHull,
         NavMesh,
+        NavGrid,
         // raw draw
         drawSprite,
         drawText,