+++ CHANGELOG.md
@@ -31,6 +31,8 @@ best friend, lajbel, can put the correct version name here
   can be used for fixed objects which do not necessarily collide in world space.
   Note that this involves additional processing as it tests outside the
   collision system, which works in world space - @mflerackers
+- Added `buildConnectivityMap()` - @mflerackers
+- Added `buildConvexHull()` - @mflerackers
 
 ## Changed